Lost streak message
The text "You lost your streak!" only shows up when the user did not have a streak to begin with. I propose changing it to say "You lost your coinflip." I moved the "your streak" part to the message where you DO lose a streak. Feel free to edit this if you do not like the wording.